Moved needed stuff from interface defaults to the builder
